03-dev/007-DevMate/requirements.txt
# TraeMate 依赖包
# 核心依赖
pyautogui>=0.9.54
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0
psutil>=5.8.0
pystray>=0.19.0

# GUI相关
# tkinter 通常随Python安装，无需单独安装

# 打包相关
PyInstaller>=4.5.0

# 开发和测试相关
pytest>=6.0.0
pytest-cov>=2.12.0
